[PATCH] Add Fuse backup path parameter
Add Fuse backup path parameter so that we can specify the local folder
instead of default path "/".

Suppose the local folder we want to sync is /tmp/backup/:

./connected test/ /tmp/backup

Note that there is no "/" at the end of backup path. By doing this
we can see all the files under /tmp/backup/ appear under test/ and
all the files of the two folders are synced.

Signed-off-by: Andiry Xu <jix024@cs.ucsd.edu>